--- apps/io-services-cms-webapp/src/watchers/index.ts
@@ -1,1 +1,0 @@
-export { handler as onServiceLifecycleChangeHandler } from "./on-services-lifecycles-change";
